requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
aiohttp>=3.10,<4
python-multipart>=0.0.9
pywebview>=5.0
playwright>=1.40
